backend/requirements.txt
# Add to backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.12.0
requests==2.31.0
python-dotenv==1.0.0
arxiv==2.1.3
pydantic==2.5.0
beautifulsoup4==4.12.2
aiohttp==3.9.1
jupyter==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1